Implement shuffle command
Added a new "shuffle" command to the qtest.
The shuffle command applies Fisher-Yates algorithm to
shuffle the elements of the queue.
Ensured the correct queue manipulation and proper
handling of different queue states.
This change includes the addition of the "do_shuffle"
function and the necessary updates to
the command registration system.

Change-Id: I116c76132f043134ce4e38e3da59eaf94e3f84fe